+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 Change Log
 ==========
 
-v.Next
+v0.2.0
 ------
 
+- Adds some system metrics from /machine/system_info 
+- Adds some Disk Storage mertics from /server/files/directory
+- Add Job Queue metric from /server/job_queue/status queue length
 - Added build targets for different platforms
 
 v0.1.1